#  add bpftrace scripts to track performance and behavior of SOF functions in the kernel