--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 logs/
 tools/__pycache__/
+kernel_tracing/metric_evaluation/__pycache__/