--- a/kernel_tracing/bpftrace_scripts/hda_irq_source_counts.bt
+++ b/kernel_tracing/bpftrace_scripts/hda_irq_source_counts.bt
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright(c) 2022 Intel Corporation. All rights reserved.
+
+// This keeps count of the irqs from each source
+
+kprobe:hda_dsp_interrupt_thread {
+  @total = count();
+}
+
+tracepoint:sof_intel:sof_intel_hda_irq {
+  @counts[str(args->source)] = count();
+}