--- a/kernel_tracing/bpftrace_scripts/hda_irq_source_deltas.bt
+++ b/kernel_tracing/bpftrace_scripts/hda_irq_source_deltas.bt
@@ -1,0 +1,15 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright(c) 2022 Intel Corporation. All rights reserved.
+
+// This collects the deltas between irqs from each source
+
+tracepoint:sof_intel:sof_intel_hda_irq {
+  if (@times[str(args->source)]) {
+    @delta_usecs[str(args->source)] = hist((nsecs - @times[str(args->source)]) / 1000);
+  }
+  @times[str(args->source)] = nsecs;
+}
+
+END {
+  clear(@times)
+}