--- a/kernel_tracing/bpftrace_scripts/ipc_tx_time.bt
+++ b/kernel_tracing/bpftrace_scripts/ipc_tx_time.bt
@@ -1,0 +1,30 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright(c) 2022 Intel Corporation. All rights reserved.
+
+// This tracks the time it takes for ipc communications (sof_ipc_tx_message execution time)
+// It prints out timings as they happen, and will generate a histogram
+// when you stop the script with ctrl-c
+
+kprobe:sof_ipc_tx_message {
+  if (@start != 0 ) {
+    printf("ERROR: overlapping tx");
+    exit();
+  }
+  @start = nsecs;
+  // We could get also the name of the device and other properties from the args,
+  // but we'd need to enable DEBUG_INFO and DEBUG_INFO_BTF to get the types
+  // It would allow us to sort timings by device for example
+  printf("tx started\n");
+}
+
+kretprobe:sof_ipc_tx_message {
+  printf("tx finished, took %d nsecs\n", nsecs - @start);
+  @usecs = hist((nsecs - @start) / 1000);
+  @avg = avg((nsecs - @start) / 1000);
+  @start = 0;
+}
+
+END {
+  // Clean up variables that shouldn't be printed
+  delete(@start);
+}