--- a/kernel_tracing/metric_evaluation/sample_shell.sh
+++ b/kernel_tracing/metric_evaluation/sample_shell.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+SOF_LOGGING=none TPLG=/lib/firmware/intel/sof-tplg/sof-hda-generic-4ch.tplg ../../test-case/check-playback.sh -l1
+sleep 5