--- a/kernel_tracing/readme.md
+++ b/kernel_tracing/readme.md
@@ -9,4 +9,6 @@ output data either as they run or when stopped with ctrl-c. To use them, you'll 
 kconfig/bpf-defconfig in your kernel config, and install `bpftrace` via your package manager. Then,
 just run them with `bpftrace [path to script]`.
 
-Soon, kernel tracing will be integrated into the CI process to detect performance regressions using tools that allow for collection and evaluation of BPF programs' output.
+Soon, kernel tracing will be integrated into the CI process to detect performance regressions. The
+`kernel_tracing/metric_evaluation` folder contains a script that collects bpftrace data and
+evaluates conditions with pytest, which would be a good candidate for CI integration.