--- a/python-deps.txt
+++ b/python-deps.txt
@@ -3,3 +3,4 @@ python3-scipy
 pylint
 python3-graphviz
 python3-construct=2.8.*
+python3-pytest