@marc-hb 2022-03-11
> it's better to squash all the commits together and do a git push --force to have a single commit, no?

@nklayman has just used Github the way it was designed to be used, not the way we use it.

Github tolerates but does not really support rewriting history and force pushes. Github is optimized for a single commit per pull request + fixup commits just like @nklayman did. For instance try the "View changes" button above: this time it works! Same for many other Github features.

More info at https://github.com/zephyrproject-rtos/zephyr/issues/39194

Note Github has an option to squash all commits right before merging. Not sure what is the final commit message then.