@marc-hb 2022-03-12
_As long as the goal is to have a single commit in the end_, I totally agree the process is much better and the end result is the same. It is the way Github is designed to be used after all!

The problem is purely about **consistency**: as Github PRs do not support reviewing multiple commits "natively", this means learning how to use Github TWICE, in two different ways. One for reviewing a single commit (in the end), and a second workflow for reviewing multiple commits (in the end).

I don't mind at all but I've been the "local git guru" in many teams for many years so I'm not representative. I've seen many, many people struggling with ONE way to use git or Github.