@nklayman 2022-04-14
@marc-hb Thanks for the review! Almost all your comments should be addressed now. The no-dependency test option is a bit tricky because the script can't really work without bpftrace. If you're okay with having that a dependency I could make a very simple case, otherwise I could try aliasing `bpftrace` to echo out some sample output so there's truly no dependencies. Let me know which one you prefer.