@marc-hb 2022-04-14
> The no-dependency test option is a bit tricky because the script can't really work without bpftrace

The dependency on bpftrace user space is not a problem because it's as simple as `apt install bpftrace`. The annoying part is the kernel part. Reconfiguring and rebooting a kernel before submitting a fix for a minor python warning is not convenient. So I'm just wondering if it would be possible to have some kind of "bpftrace no-op" example / test case for this = anything that does not require a new kernel. If bpftrace has some hard kernel dependencies that are typically not available in stock kernels provided by distributions then too bad forget it. Maybe we could add some sort of "mock bpftrace" later.

(I've not looked at the new version yet)