@marc-hb 2022-04-15
> I think we have parametrization already but maybe @pytest.mark.parametrize is more "pythonic", flexible and generally better way to do it

I just remembered reading somewhere complaining that `@pytest.mark.parametrize` is static and can't be used after reading a config file. Vague and second-hand information, sorry for any noise.