@nklayman 2022-04-18
@marc-hb latest push includes a `no_dependencies.json` that should work with just bpftrace installed. Also fixes edge cases where the spec file arg doesn't start with a `./` and when the bpftrace script doesn't exit in time.