@marc-hb 2022-04-19
Nice try but:
```
collecting ... modprobe: FATAL: Module kheaders not found in directory /lib/modules/9.9.9-rc9softest
$ ['sleep', '1']
BPF tracing finished
```

Also, not running as root still hangs:
```
collecting ... ERROR: bpftrace currently only supports running as the root user.
  ... hangs ...
```

```
$ ps f

    987 pts/0    Ss     0:00 -bash
   1161 pts/0    R+     0:39  \_ python3 ./kernel_tracing/metric_evaluation/test.py kernel_tracing/metric_evaluation/no_dependencies.json
   1163 pts/0    Z+     0:00      \_ [bpftrace] <defunct>
```

BTW only bpftrace should run as root, should be started as `sudo bpftrace`.

Again all this can be fixed later.